backend_fastAPI/requirements.txt
matplotlib==3.8.2
numpy==1.26.4
Pillow
tensorflow==2.15.0
fastapi
pydantic